test: split 2 unrelated nuts
